+++ tests/tests_pytorch/loggers/test_all.py
@@ -105,7 +105,9 @@ def test_loggers_fit_test_all(logger_class, mlflow_mock, wandb_mock, comet_mock,
 
     if logger_class == CometLogger:
         logger.experiment.id = "foo"
-        logger.experiment.project_name = "bar"
+        logger._comet_config.offline_directory = None
+        logger._project_name = "bar"
+        logger.experiment.get_key.return_value = "SOME_KEY"
 
     if logger_class == NeptuneLogger:
         logger._retrieve_run_data = Mock()
@@ -292,7 +294,9 @@ def test_logger_with_prefix_all(mlflow_mock, wandb_mock, comet_mock, neptune_moc
     _patch_comet_atexit(monkeypatch)
     logger = _instantiate_logger(CometLogger, save_dir=tmp_path, prefix=prefix)
     logger.log_metrics({"test": 1.0}, step=0)
-    logger.experiment.log_metrics.assert_called_once_with({"tmp-test": 1.0}, epoch=None, step=0)
+    logger.experiment.__internal_api__log_metrics__.assert_called_once_with(
+        {"test": 1.0}, epoch=None, step=0, prefix=prefix, framework="pytorch-lightning"
+    )
 
     # MLflow
     Metric = mlflow_mock.entities.Metric